+++ components/com_crowdfunding/views/project/tmpl/funding.php
@@ -10,13 +10,12 @@
 // no direct access
 defined('_JEXEC') or die;
 
-if (strcmp('five_steps', $this->wizardType) == 0) {
+if (strcmp('five_steps', $this->wizardType) === 0) {
     $layout      = new JLayoutFile('project_wizard');
 } else {
     $layout      = new JLayoutFile('project_wizard_six_steps');
 }
 echo $layout->render($this->layoutData);
-
 ?>
 <div class="row">
     <div class="col-md-12">
@@ -24,7 +23,7 @@ echo $layout->render($this->layoutData);
 
             <div class="row">
                 <div class="col-md-2"><?php echo $this->form->getLabel('goal'); ?></div>
-                <div class="col-md-4">
+                <div class="col-md-10">
                     <?php echo $this->form->getInput('goal'); ?>
                     <?php if(!empty($this->maxAmount)) {?>
                     <span class="help-block"><?php echo JText::sprintf('COM_CROWDFUNDING_MINIMUM_MAXIMUM_AMOUNT', $this->amount->setValue($this->minAmount)->formatCurrency(), $this->amount->setValue($this->maxAmount)->formatCurrency());?></span>
@@ -34,21 +33,21 @@ echo $layout->render($this->layoutData);
                 </div>
             </div>
 
-            <div class="row">
+            <div class="row mt-20 bt-dashed-thin pt-20">
                 <div class="col-md-2"><?php echo $this->form->getLabel('funding_type'); ?></div>
                 <div class="col-md-10">
                     <?php echo $this->form->getInput('funding_type'); ?>
                 </div>
             </div>
         
-        	<div class="row">
+        	<div class="row mt-20 bt-dashed-thin pt-20">
                 <div class="col-md-2">
-                	<label title="<?php echo JHtml::tooltipText(JText::_('COM_CROWDFUNDING_FIELD_FUNDING_DURATION_DESC'));?>" class="hasTooltip" for="jform_funding_duration_type" id="jform_funding_duration_type-lbl">
+                	<label <?php echo JHtml::_('Prism.ui.popoverText', JText::_('COM_CROWDFUNDING_FIELD_FUNDING_DURATION_DESC'),JText::_('COM_CROWDFUNDING_FIELD_FUNDING_DURATION'));?> class="hasPopover" for="jform_funding_duration_type" id="jform_funding_duration_type-lbl">
                 	<?php echo JText::_('COM_CROWDFUNDING_FIELD_FUNDING_DURATION');?><span class="star">&nbsp;*</span>
                 	</label>
                 </div>
                 
-                <div class="col-md-4">
+                <div class="col-md-10">
                     <?php if (!$this->fundingDuration or (strcmp('days', $this->fundingDuration) === 0)) {?>
                         <input type="radio" value="days" name="jform[funding_duration_type]" id="js-funding-duration-days" <?php echo $this->checkedDays;?>/>
                         <?php echo $this->form->getLabel('funding_days'); ?>